fix: move functions to avoid circular dependencies
